LayoutTests/fast/css/parsing-relative-color-syntax-expected.txt

Test the parsing of CSS Color 5 Relative Color Syntax.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


rgb(from ...)
PASS computedStyle("background-color", "rgb(from rebeccapurple r g b)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g b / alpha)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g b / alpha)") is "rgba(26, 51, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from hsl(120deg 20% 50% / .5) r g b / alpha)") is "rgba(102, 153, 102, 0.5)"
PASS computedStyle("background-color", "rgb(from rgb(from rebeccapurple r g b) r g b)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "rgb(from color(display-p3 0 1 0) r g b / alpha)") is "rgb(0, 255, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 0 0 0)") is "rgb(0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 0 0 0 / 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 0 g b / alpha)") is "rgb(0, 51, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r 0 b / alpha)") is "rgb(102, 0, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g 0 / alpha)") is "rgb(102, 51, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g b / 0)") is "rgba(102, 51, 153, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) 0 g b / alpha)") is "rgba(0, 51, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r 0 b / alpha)") is "rgba(26, 0, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g 0 / alpha)") is "rgba(26, 51, 0, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g b / 0)") is "rgba(26, 51, 77, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 25 g b / alpha)") is "rgb(25, 51, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r 25 b / alpha)") is "rgb(102, 25, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g 25 / alpha)") is "rgb(102, 51, 25)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g b / .25)") is "rgba(102, 51, 153, 0.25)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) 25 g b / alpha)") is "rgba(25, 51, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r 25 b / alpha)") is "rgba(26, 25, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g 25 / alpha)") is "rgba(26, 51, 25, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g b / .25)") is "rgba(26, 51, 77, 0.25)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 25% g b / alpha)") is "rgb(64, 51, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r 25% b / alpha)") is "rgb(102, 64, 153)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g 25% / alpha)") is "rgb(102, 51, 64)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g b / 25%)") is "rgba(102, 51, 153, 0.25)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) 25% g b / alpha)") is "rgba(64, 51, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r 25% b / alpha)") is "rgba(26, 64, 77, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g 25% / alpha)") is "rgba(26, 51, 64, 0.4)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g b / 25%)") is "rgba(26, 51, 77, 0.25)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 25 g b / 25%)") is "rgba(25, 51, 153, 0.25)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r 25 b / 25%)") is "rgba(102, 25, 153, 0.25)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r g 25 / 25%)") is "rgba(102, 51, 25, 0.25)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) 25 g b / 25%)") is "rgba(25, 51, 77, 0.25)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r 25 b / 25%)") is "rgba(26, 25, 77, 0.25)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r g 25 / 25%)") is "rgba(26, 51, 25, 0.25)"
PASS computedStyle("background-color", "rgb(from rebeccapurple g b r)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple b alpha r / g)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r r r / r)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple alpha alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) g b r)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) b alpha r / g)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r r r / r)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) alpha alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple r 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple 0% 10 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) r 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rgb(10%, 20%, 30%, 40%) 0% 10 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple red g b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple l g b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "rgb(from rebeccapurple h g b)") is "rgba(0, 0, 0, 0)"

hsl(from ...)
PASS computedStyle("background-color", "hsl(from rebeccapurple h s l)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h s l / alpha)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h s l / alpha)") is "rgba(26, 51, 77, 0.4)"
PASS computedStyle("background-color", "hsl(from hsl(120deg 20% 50% / .5) h s l / alpha)") is "rgba(102, 153, 102, 0.5)"
PASS computedStyle("background-color", "hsl(from hsl(from rebeccapurple h s l) h s l)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "hsl(from color(display-p3 0 1 0) h s l / alpha)") is "rgb(0, 255, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 0 0% 0%)") is "rgb(0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 0deg 0% 0%)") is "rgb(0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 0 0% 0% / 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 0deg 0% 0% / 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 0 s l / alpha)") is "rgb(153, 51, 51)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 0deg s l / alpha)") is "rgb(153, 51, 51)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h 0% l / alpha)") is "rgb(102, 102, 102)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h s 0% / alpha)") is "rgb(0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h s l / 0)") is "rgba(102, 51, 153, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) 0 s l / alpha)") is "rgba(77, 26, 26, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) 0deg s l / alpha)") is "rgba(77, 26, 26, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h 0% l / alpha)") is "rgba(52, 52, 52, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h s 0% / alpha)") is "rgba(0, 0, 0, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h s l / 0)") is "rgba(26, 51, 77, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 25 s l / alpha)") is "rgb(153, 94, 51)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 25deg s l / alpha)") is "rgb(153, 94, 51)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h 25% l / alpha)") is "rgb(102, 77, 128)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h s 25% / alpha)") is "rgb(64, 32, 96)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h s l / .25)") is "rgba(102, 51, 153, 0.25)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) 25 s l / alpha)") is "rgba(77, 47, 26, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) 25deg s l / alpha)") is "rgba(77, 47, 26, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h 25% l / alpha)") is "rgba(39, 51, 64, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h s 25% / alpha)") is "rgba(32, 63, 95, 0.4)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h s l / .25)") is "rgba(26, 51, 77, 0.25)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h l s)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h alpha l / s)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h l l / l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h l s)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h alpha l / s)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h l l / l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple 10% s l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) h 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rgb(10%, 20%, 30%, 40%) 10% s l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple hue s l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple x s l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hsl(from rebeccapurple h g b)") is "rgba(0, 0, 0, 0)"

hwb(from ...)
PASS computedStyle("background-color", "hwb(from rebeccapurple h w b)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h w b / alpha)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h w b / alpha)") is "rgba(26, 51, 77, 0.4)"
PASS computedStyle("background-color", "hwb(from hsl(120deg 20% 50% / .5) h w b / alpha)") is "rgba(102, 153, 102, 0.5)"
PASS computedStyle("background-color", "hwb(from hwb(from rebeccapurple h w b) h w b)") is "rgb(102, 51, 153)"
PASS computedStyle("background-color", "hwb(from color(display-p3 0 1 0) h w b / alpha)") is "rgb(0, 255, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 0 0% 0%)") is "rgb(255, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 0deg 0% 0%)") is "rgb(255, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 0 0% 0% / 0)") is "rgba(255, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 0deg 0% 0% / 0)") is "rgba(255, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 0 w b / alpha)") is "rgb(153, 51, 51)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 0deg w b / alpha)") is "rgb(153, 51, 51)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h 0% b / alpha)") is "rgb(77, 0, 153)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h w 0% / alpha)") is "rgb(153, 51, 255)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h w b / 0)") is "rgba(102, 51, 153, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) 0 w b / alpha)") is "rgba(77, 26, 26, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) 0deg w b / alpha)") is "rgba(77, 26, 26, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h 0% b / alpha)") is "rgba(0, 38, 77, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h w 0% / alpha)") is "rgba(26, 138, 255, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h w b / 0)") is "rgba(26, 51, 77, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 25 w b / alpha)") is "rgb(153, 94, 51)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 25deg w b / alpha)") is "rgb(153, 94, 51)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h 25% b / alpha)") is "rgb(108, 64, 153)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h w 25% / alpha)") is "rgb(121, 51, 191)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h w b / .25)") is "rgba(102, 51, 153, 0.25)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) 25 w b / alpha)") is "rgba(77, 47, 26, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) 25deg w b / alpha)") is "rgba(77, 47, 26, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h 25% b / alpha)") is "rgba(64, 70, 77, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h w 25% / alpha)") is "rgba(26, 107, 191, 0.4)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h w b / .25)") is "rgba(26, 51, 77, 0.25)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h b w)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h alpha w / b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h w w / w)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h b w)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h alpha w / b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h w w / w)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple 10% w b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) h 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rgb(10%, 20%, 30%, 40%) 10% w b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple hue w b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple x w b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "hwb(from rebeccapurple h g b)") is "rgba(0, 0, 0, 0)"

lab(from ...)
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a b)") is "lab(25% 20 50)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a b / alpha)") is "lab(25% 20 50)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l a b / alpha)") is "lab(25% 20 50 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(from lab(25% 20 50) l a b) l a b)") is "lab(25% 20 50)"
PASS computedStyle("background-color", "lab(from color(display-p3 0 0 0) l a b / alpha)") is "lab(0% 0 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) 0% 0 0)") is "lab(0% 0 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) 0% 0 0 / 0)") is "lab(0% 0 0 / 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) 0% a b / alpha)") is "lab(0% 20 50)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l 0 b / alpha)") is "lab(25% 0 50)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a 0 / alpha)") is "lab(25% 20 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a b / 0)") is "lab(25% 20 50 / 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) 0% a b / alpha)") is "lab(0% 20 50 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l 0 b / alpha)") is "lab(25% 0 50 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l a 0 / alpha)") is "lab(25% 20 0 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l a b / 0)") is "lab(25% 20 50 / 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) 35% a b / alpha)") is "lab(35% 20 50)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l 35 b / alpha)") is "lab(25% 35 50)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a 35 / alpha)") is "lab(25% 20 35)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a b / .35)") is "lab(25% 20 50 / 0.35)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) 35% a b / alpha)") is "lab(35% 20 50 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l 35 b / alpha)") is "lab(25% 35 50 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l a 35 / alpha)") is "lab(25% 20 35 / 0.4)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l a b / .35)") is "lab(25% 20 50 / 0.35)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l b a)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l alpha a / b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l a a / a)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l b a)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l alpha a / b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l a a / a)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l alpha alpha / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) l 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) 10 a b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l 10% 10)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) l 10 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50 / 40%) 10 a b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) lightness a b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) x a b)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lab(from lab(25% 20 50) h g b)") is "rgba(0, 0, 0, 0)"

lch(from ...)
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c h)") is "lch(70% 45 30)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c h / alpha)") is "lch(70% 45 30)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c h/ alpha)") is "lch(70% 45 30 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(from lch(70% 45 30) l c h) l c h)") is "lch(70% 45 30)"
PASS computedStyle("background-color", "lch(from color(display-p3 0 0 0) l c h / alpha)") is "lch(0% 0 0)"
PASS computedStyle("background-color", "lch(from color(lab 70% 45 30) l c h / alpha)") is "lch(70% 54.08327 33.690067)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 0% 0 0)") is "lch(0% 0 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 0% 0 0deg)") is "lch(0% 0 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 0% 0 0 / 0)") is "lch(0% 0 0 / 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 0% 0 0deg / 0)") is "lch(0% 0 0 / 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 0% c h / alpha)") is "lch(0% 45 30)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l 0 h / alpha)") is "lch(70% 0 30)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c 0 / alpha)") is "lch(70% 45 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c 0deg / alpha)") is "lch(70% 45 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c h / 0)") is "lch(70% 45 30 / 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) 0% c h / alpha)") is "lch(0% 45 30 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l 0 h / alpha)") is "lch(70% 0 30 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c 0 / alpha)") is "lch(70% 45 0 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c 0deg / alpha)") is "lch(70% 45 0 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c h / 0)") is "lch(70% 45 30 / 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 25% c h / alpha)") is "lch(25% 45 30)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l 25 h / alpha)") is "lch(70% 25 30)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c 25 / alpha)") is "lch(70% 45 25)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c 25deg / alpha)") is "lch(70% 45 25)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c h / .25)") is "lch(70% 45 30 / 0.25)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) 25% c h / alpha)") is "lch(25% 45 30 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l 25 h / alpha)") is "lch(70% 25 30 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c 25 / alpha)") is "lch(70% 45 25 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c 25deg / alpha)") is "lch(70% 45 25 / 0.4)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c h / .25)") is "lch(70% 45 30 / 0.25)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) alpha c h / l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) alpha c h / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) alpha c h / l)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) alpha c h / alpha)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l 10% h)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l c 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) 10 c h)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l 10% h)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) l c 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30 / 40%) 10 c h)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) lightness c h)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) x c h)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "lch(from lch(70% 45 30) l g b)") is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
